[Fix] Silence another conversion warning. [Fix] Remove bogus file entry from Android and autotools makefiles.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@16433 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Sndmix.cpp
@@ -1172,7 +1172,7 @@ int CSoundFile::ProcessPitchFilterEnvelope(ModChannel &chn, int32 &period) const
 #else
 		// TODO: AMS2 envelopes behave differently when linear slides are off - emulate with 15 * (-128...127) >> 6
 		// Copy over vibrato behaviour for that?
-		const int32 range = GetType() == MOD_TYPE_AMS ? uint8_max : ENVELOPE_MAX;
+		const int32 range = GetType() == MOD_TYPE_AMS ? uint8_max : uint8(ENVELOPE_MAX);
 		int32 amp;
 		switch(GetType())
 		{
